Fix the final-review findings in the pipeline
Money path (label stage): thinking is explicitly disabled on the models
whose default is adaptive thinking, max_tokens is 256, and a 50-request
canary batch must parse at 90% before the rest of the sample is
submitted (CanaryFailed, CLI exit 5). The batch id is written to the run
log the moment a batch is created so a Ctrl-C cannot orphan a paid
batch, and `loupe label --resume-batch ID` polls an existing one. A bad
result body costs one row instead of the whole batch, the labels parquet
is written in a finally, and a run that labels under 90% of the sample
is flagged not ok (CLI exit 4). Actual cost now counts cache tokens.

Metrics: return_rate is NULL for weeks whose following week is absent
from the data (with a next_week_present flag) instead of publishing 0.0,
which the dashboard would have read as churn; a new
pseudo_user_persistence_quarterly metric shows the pseudo-user key
losing week-to-week persistence around 2024-10. Country and language
volume files publish a suppressed_or_unknown residual row per week
(8.3% of the corpus for country) instead of dropping suppressed and
NULL cells. Slices group by model family, not by dated point release.
Intent coverage is "full" only when every flattened shard has
predictions. meta.json gains complete_weeks_from/through,
token_coverage_first_week, intent_share, the classifier report fields
and a caveats list. Sums are cast to BIGINT for Arrow JS, and coverage
rates no longer read -0.0.

+++ loupe/metrics/intent_by_language.sql
@@ -1,5 +1,5 @@
 WITH top_lang AS (SELECT language FROM conversations GROUP BY language ORDER BY count(*) DESC LIMIT 10)
-SELECT c.language, i.intent, count(*) AS conversations
+SELECT c.language, i.intent, count(*)::BIGINT AS conversations
 FROM conversations c JOIN intent i USING (conv_id) JOIN top_lang USING (language)
 WHERE i.intent IS NOT NULL
 GROUP BY ALL HAVING count(*) >= {{min_cell}} ORDER BY c.language, conversations DESC
